10-30
29，把所有的红包编码加上了（加了包体回收方式字段）
解决扣款余额问题
30.订单列表分页查询优化
解决页面特殊包体显示枚举问题
页面显示优化（主要是代理商账户类型）
更新接口文档
回调：需要对接人员编写了回调接口，才能测试回调
还剩同样类型通道的包体合并筛选出所有可充值的包体（可以利用编码）

+++ weizu-parent/weizu-web-modules/weizu-channel/src/main/java/com/weizu/flowsys/web/trade/dao/impl/PurchaseDaoImpl.java
@@ -53,7 +53,7 @@ public class PurchaseDaoImpl extends DaoImpl<PurchasePo, Long> implements Purcha
 	 * @createTime:2017年6月13日 下午12:47:58
 	 */
 	@Override
-	public int countPurchase(Map<String, Object> paramsMap) {
+	public long countPurchase(Map<String, Object> paramsMap) {
 		paramsMap.put("accountType", AccountTypeEnum.DECREASE.getValue());
 		return sqlSessionTemplate.selectOne("countPurchase", paramsMap);
 	}
